move update_location from Person to Location
