+++ app/models/district.rb
@@ -47,6 +47,13 @@ class District < ActiveRecord::Base
     senators.eligible.unconvinced + [unconvinced_rep].compact
   end
 
+  def legislators
+    join_clause = 'INNER JOIN districts '\
+                  'ON (districts.id = legislators.district_id '\
+                  'OR districts.state_id = legislators.state_id)'
+    Legislator.joins(join_clause).where(districts: {id: id})
+  end
+
   def targeted?
     campaigns.active.any?
   end
